Add support to compile with VS2013 x64

+++ src/allegro/src/glyph.c
@@ -25,7 +25,7 @@
 #define DRAW_GLYPH(bits, size)                                               \
 {                                                                            \
    AL_CONST unsigned char *data = glyph->dat;                                \
-   unsigned long addr;                                                       \
+   uintptr_t addr;                                                           \
    int w = glyph->w;                                                         \
    int h = glyph->h;                                                         \
    int stride = (w+7)/8;                                                     \
